Add pypirc check
